[gs/jh] account for periods of no istio convergences
We have an SLO for istio convergences happening in a certain amount of
time.  If there are no convergences for a while, the rate of events
within the window drops to 0, leading to a division by 0, and the SLO
becomes NaN when we really want to say that we have no errors and
therefore are not failing.  So we substitute a rate of 1 for these time
periods.  The value doesn't matter as long as it's >0, because the
rate of errors at this time must be 0.

Reference: https://github.com/slok/sloth/issues/231

Also, fix the convergence target from 0.1 to 0.5, as the exercise
requires.